+++ src/main/java/ru/hse/plugin/metrics/editor/AllCharCounter.java
@@ -1,11 +1,12 @@
-package ru.hse.plugin.metrics;
+package ru.hse.plugin.metrics.editor;
 
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiFile;
 import org.jetbrains.annotations.NotNull;
 import ru.hse.plugin.localstat.AllSymbolStatistics;
-import ru.hse.plugin.metrics.component.MetricJComponentWrapper;
+import ru.hse.plugin.metrics.Metric;
+import ru.hse.plugin.metrics.commons.component.MetricJComponentWrapper;
 import ru.hse.plugin.util.PluginConstants;
 
 import javax.swing.*;
