+++ src/main/java/ru/hse/plugin/metrics/project/ProjectOpensNumber.java
@@ -1,9 +1,10 @@
-package ru.hse.plugin.metrics;
+package ru.hse.plugin.metrics.project;
 
 import com.intellij.openapi.project.Project;
 import org.jetbrains.annotations.NotNull;
-import ru.hse.plugin.metrics.component.CounterJComponentWrapper;
-import ru.hse.plugin.metrics.component.MetricJComponentWrapper;
+import ru.hse.plugin.metrics.Metric;
+import ru.hse.plugin.metrics.commons.component.CounterJComponentWrapper;
+import ru.hse.plugin.metrics.commons.component.MetricJComponentWrapper;
 import ru.hse.plugin.util.PluginConstants;
 
 public class ProjectOpensNumber extends Metric {
